read data
added read;
folds and types (train, eval) can be specified
for each fold there is a dictionary containing the fold as key a dictionary of the corresponding fold types as value. in each fold type dictionary there is a another dictionary containing the label as a key and all the read numpy files as (matrix) value.

i tried merging the matrizes if each label but it's just too much data (would give you folds*foldTypes*label(60*500*amount of numpy files) matrizes...